+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+2021.08
+-------------------------------------------------------------------------------
+- completely removed the aldryn stack (aldryn-addons, aldryn-django, aldryn-djangocms)
+- django CMS 3.9
+- abandon custom docker images
+- new docker-based build process
+- proper whitenoise static files integration
+- proper frontend webpack-dev-server hot reload setup
+- fix HMR (webpack dev server hot reloading)
+
+
 2021.06
 -------------------------------------------------------------------------------